# [Perf] 6.5x faster LiteLLM Python SDK Completion  