--- a/litellm/utils.py
+++ b/litellm/utils.py
@@ -3088,6 +3088,7 @@ def pre_process_non_default_params(
     model: str,
     remove_sensitive_keys: bool = False,
     add_provider_specific_params: bool = False,
+    provider_config: Optional[BaseConfig] = None,
 ) -> dict:
     """
     Pre-process non-default params to a standardized format
@@ -3102,14 +3103,6 @@ def pre_process_non_default_params(
         default_param_values=DEFAULT_CHAT_COMPLETION_PARAM_VALUES,
         additional_endpoint_specific_params=["messages"],
     )
-
-    provider_config: Optional[BaseConfig] = None
-    if custom_llm_provider is not None and custom_llm_provider in [
-        provider.value for provider in LlmProviders
-    ]:
-        provider_config = ProviderConfigManager.get_provider_chat_config(
-            model=model, provider=LlmProviders(custom_llm_provider)
-        )
 
     if "response_format" in non_default_params:
         if provider_config is not None: