## [Perf] 6.5x faster LiteLLM Python SDK Completion  

Optimizes code in litellm.completion to reduce duplicate computation of the llm provider config 

| Metric | Before | After (Run 1) | After (Run 2) | Improvement (Run 1) | Improvement (Run 2) |
|--------|--------|---------------|---------------|-------------------|-------------------|
| **preprocess_params Total Time** | 25.78s | 1.89s | 0.76s | **13.6x faster** | **33.9x faster** |
| **preprocess_params Avg per Call** | 2.58ms | 0.19ms | 0.08ms | **13.6x faster** | **32.3x faster** |
| **preprocess_params % of Total** | 15.8% | 4.5% | 3.0% | -71% | -81% |
| **Overall Total Time** | 163.08s | 41.82s | 25.12s | **3.9x faster** | **6.5x faster** |

<!-- e.g. "Implement user authentication feature" -->

## Relevant issues

<!-- e.g. "Fixes #000" -->

## Pre-Submission checklist

**Please complete all items before asking a LiteLLM maintainer to review your PR**

- [ ] I have Added testing in the [`tests/litellm/`](https://github.com/BerriAI/litellm/tree/main/tests/litellm) directory, **Adding at least 1 test is a hard requirement** - [see details](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] I have added a screenshot of my new test passing locally 
- [ ] My PR passes all unit tests on [`make test-unit`](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] My PR's scope is as isolated as possible, it only solves 1 specific problem


## Type

<!-- Select the type of Pull Request -->
<!-- Keep only the necessary ones -->

🚄 Infrastructure
✅ Test

## Changes